terrain added + title screen animation added (read desc)
finished with the main starting area, the place where you will talk to the bug to get in to the actual village

also, began terrain for the searchable area, tree parkour etc. For now, i'm only doing one area because of time.

added title screen UI as placeholders as well as an animation that moves the scene view